pyproject.toml
[project]
name = "pdf-to-docx-fitz"
version = "0.1.0"
description = "Add your description here"
authors = [
    { name = "vmerlyn", email = "vibeesh.bose@merlynsoftware.com" }
]
dependencies = [
    "fastapi>=0.116.1",
    "uvicorn>=0.35.0",
    "python-docx>=1.2.0",
    "python-multipart>=0.0.20",
    "pymupdf>=1.26.3",
    "routes>=2.5.1",
]
readme = "README.md"
requires-python = ">= 3.8"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/pdf_to_docx_fitz"]
